+++ src/app/community/packages2/[ecosystem]/[...nameAndVersion]/_components/tabs/vulnerabilities-tab.tsx
