+++ src/app/community/packages2/[ecosystem]/[...nameAndVersion]/page.tsx
@@ -5,10 +5,10 @@ import { Metadata } from "next";
 import { Tabs, TabsList, TabsTrigger, TabsContent } from "@/components/ui/tabs";
 import PackageHeader from "./_components/package-header";
 import StatsCards from "./_components/stats-cards";
-import AnalysisTab from "./_components/analysis-tab";
+import AnalysisTab from "./_components/tabs/analysis-tab";
 import { Suspense } from "react";
-import VulnerabilitiesTab from "./_components/vulnerabilities-tab";
-import VersionsTab from "./_components/versions-tab";
+import VulnerabilitiesTab from "./_components/tabs/vulnerabilities-tab";
+import VersionsTab from "./_components/tabs/versions-tab";
 
 export async function generateMetadata({
   params,
